Clean-up imports and code structure. Isolated long sleeping (>40d) into standalone function. Cleaned printing and logging statements. Deleted `embeds.py`, moved all pagination code into `pagination.py`.
